backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6

# AI / LLM
anthropic==0.42.0
voyageai==0.3.2

# Task Queue
celery[redis]==5.4.0
redis==5.2.1

# Document Parsing
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
pdf2image==1.17.0

# OCR
pytesseract==0.3.13
Pillow==11.1.0

# Document Generation
docxtpl==0.18.0

# Google Sheets
gspread==6.1.4
google-auth==2.37.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Storage
minio==7.2.12

# Configuration
pydantic-settings==2.7.1

# HTTP
httpx==0.28.1

# Utilities
tenacity==9.0.0
python-dateutil==2.9.0
numpy==2.2.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
